introduce useFocusButtonTrap
