+++ x-pack/platform/plugins/shared/cases/public/components/actions/tags/edit_tags_flyout.tsx
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useCallback, useState } from 'react';
 import { css } from '@emotion/react';
 import {
   EuiButton,
@@ -16,7 +16,6 @@ import {
   EuiFlyoutBody,
   EuiFlyoutFooter,
   EuiFlyoutHeader,
-  type EuiFocusTrapProps,
   euiFullHeight,
   EuiLoadingSpinner,
   EuiText,
@@ -28,6 +27,7 @@ import { useGetTags } from '../../../containers/use_get_tags';
 import { EditTagsSelectable } from './edit_tags_selectable';
 import * as i18n from './translations';
 import type { ItemsSelectionState } from '../types';
+import { useFocusButtonTrap } from '../../use_focus_button';
 
 interface Props {
   selectedCases: CasesUI;
@@ -58,19 +58,7 @@ const EditTagsFlyoutComponent: React.FC<Props> = ({
   });
 
   const onSave = useCallback(() => onSaveTags(tagsSelection), [onSaveTags, tagsSelection]);
-
-  const focusTrapProps: Pick<EuiFocusTrapProps, 'returnFocus'> = useMemo(
-    () => ({
-      returnFocus() {
-        if (focusButtonRef && 'current' in focusButtonRef && focusButtonRef.current) {
-          focusButtonRef.current.focus();
-          return false;
-        }
-        return true;
-      },
-    }),
-    [focusButtonRef]
-  );
+  const focusTrapProps = useFocusButtonTrap(focusButtonRef);
 
   const headerSubtitle =
     selectedCases.length > 1 ? i18n.SELECTED_CASES(selectedCases.length) : selectedCases[0].title;
